pyproject.toml
[project]
name = "xianyu-fish-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=5.4",
]
